+++ src/main/java/com/kgu/studywithme/global/annotation/aop/CheckMemberIdentityAspect.java
@@ -1,4 +1,4 @@
-package com.kgu.studywithme.global.annotation;
+package com.kgu.studywithme.global.annotation.aop;
 
 import com.kgu.studywithme.auth.exception.AuthErrorCode;
 import com.kgu.studywithme.global.exception.StudyWithMeException;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 @Aspect
 @Component
 public class CheckMemberIdentityAspect {
-    @Before("@annotation(CheckMemberIdentity) && args(payloadId, memberId, ..)")
+    @Before("@annotation(com.kgu.studywithme.global.annotation.aop.CheckMemberIdentity) && args(payloadId, memberId, ..)")
     public void checkPayloadId(Long payloadId, Long memberId) {
         if (isAnonymousMember(payloadId, memberId)) {
             throw StudyWithMeException.type(AuthErrorCode.INVALID_PERMISSION);
