+++ src/main/java/com/kgu/studywithme/study/controller/notice/StudyNoticeApiController.java
@@ -1,7 +1,7 @@
 package com.kgu.studywithme.study.controller.notice;
 
 import com.kgu.studywithme.auth.utils.ExtractPayload;
-import com.kgu.studywithme.global.annotation.CheckStudyHost;
+import com.kgu.studywithme.global.annotation.aop.CheckStudyHost;
 import com.kgu.studywithme.study.controller.dto.request.NoticeRequest;
 import com.kgu.studywithme.study.service.notice.NoticeService;
 import lombok.RequiredArgsConstructor;
